refactor: convert firebaseconfig to .ts file
